pyproject.toml
[tool.poetry]
name = "computer vision (scd24)"
version = "0.1.0"
description = ""
authors = ["Daniel Amaral <silva.daniel86@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = {version = "^2.5.1+cu124", source = "pytorch"}
torchvision = {version = "^0.20.1+cu124", source = "pytorch"}
numpy = "^2.1.3"
matplotlib = "^3.9.2"
pycocotools = "^2.0.8"
tqdm = "^4.67.0"
pillow = "^11.0.0"
albumentations = "^1.4.21"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
fiftyone = "^1.0.2"
torchmetrics = "^1.6.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]